functions/requirements.txt
firebase_functions
firebase-admin
google-genai
requests
tenacity
pdfplumber
pypdf